apps/api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
PyMuPDF==1.23.8
WeasyPrint==60.2
jinja2==3.1.2
pydantic==2.5.3
google-generativeai==0.3.2  # For Gemini API
python-dotenv==1.0.0       # For loading env vars locally
